update api path
