+++ Gemfile
@@ -19,6 +19,7 @@ gem 'yaml_db'
 group :development, :test do
   gem 'debug', platforms: %i[mri mingw x64_mingw]
 end
+gem 'rswag'
 group :development do
   gem 'web-console'
 end
