+++ app/controllers/api/v1/posts_controller.rb
@@ -1,8 +1,7 @@
 module Api
   module V1
     class PostsController < ApplicationController
-      before_action :authenticate_user!
-      before_action :set_post, only: [:show]
+      skip_before_action :authenticate_user!
 
       def index
         user = User.find(params[:user_id])
@@ -11,14 +10,29 @@ module Api
         render json: posts
       end
 
-      def show
-        render json: @post
+      def comments
+        post = Post.find(params[:id])
+        comments = post.comments
+
+        render json: comments
+      end
+
+      def add_comment
+        @post = Post.find(params[:id])
+        comment = @post.comments.new(comment_params)
+        comment.user = current_user
+
+        if comment.save
+          render json: comment, status: :created
+        else
+          render json: { errors: comment.errors.full_messages }, status: :unprocessable_entity
+        end
       end
 
       private
 
-      def set_post
-        @post = Post.find(params[:id])
+      def comment_params
+        params.require(:comment).permit(:text)
       end
     end
   end
