+++ config/routes.rb
@@ -9,14 +9,19 @@ Rails.application.routes.draw do
     end
   end
 
-  namespace :api, defaults:{ format: :json } do
+  namespace :api, defaults: { format: :json } do
     namespace :v1 do
       resources :users, only: [:index, :show] do
         resources :posts, only: [:index, :show] do
-          resources :comments, only: [:index, :create]
+          member do
+            get :comments
+            post :add_comment
+          end
         end
       end
     end
   end
+  
+  
   
 end
